feat: ASR timeouts accept sub-second decimals (NUMERIC + float64)
Schema migration 0010 widens asr_silence_timeout_sec /
asr_speech_timeout_sec / asr_speech_max_sec from INT to NUMERIC(5,2)
and drops the implicit INT-min CHECK so tuning values like 0.5, 1.2,
12.75 store cleanly. PG-named CHECK constraints are discovered via
pg_constraint and dropped dynamically — robust to the unstable suffix
auto-naming would otherwise produce.

Go side flips ASRSilenceTimeoutSec / SpeechTimeoutSec / SpeechMaxSec
from int to float64 across BotConfig, BotWriteInput, SeedBotInput,
the API DTO, and the seed bootstrap in main. pipeline.session_runner
multiplies by float64(time.Second) so the decimal portion isn't
truncated by the implicit float→int64 cast that `time.Duration(x) *
time.Second` would do otherwise. web_voice.go converts the same
float seconds to int ms when handing off to asr.StreamOpts.

botSelectCols casts each NUMERIC column to FLOAT8 so pgx scans
directly into the float64 struct fields instead of returning
pgtype.Numeric.

+++ internal/api/web_voice.go
@@ -273,13 +273,13 @@ var errClientClosed = errors.New("client closed")
 // asr_partial frames for UI animation).
 func (s *voiceState) listen(ctx context.Context, audioCh <-chan []byte) (string, error) {
 	stream, err := s.asrCli.StartStream(ctx, asr.StreamOpts{
-		ConversationID:  "web-" + s.sess.ID.String(),
-		SampleRate:      s.asrRate,
-		Channels:        1,
-		SingleSentence:  s.bot.ASRSingleSentence,
-		SilenceTimeoutMs: s.bot.ASRSilenceTimeoutSec * 1000,
-		SpeechTimeoutMs:  s.bot.ASRSpeechTimeoutSec * 1000,
-		SpeechMaxMs:      s.bot.ASRSpeechMaxSec * 1000,
+		ConversationID:   "web-" + s.sess.ID.String(),
+		SampleRate:       s.asrRate,
+		Channels:         1,
+		SingleSentence:   s.bot.ASRSingleSentence,
+		SilenceTimeoutMs: int(s.bot.ASRSilenceTimeoutSec * 1000),
+		SpeechTimeoutMs:  int(s.bot.ASRSpeechTimeoutSec * 1000),
+		SpeechMaxMs:      int(s.bot.ASRSpeechMaxSec * 1000),
 	})
 	if err != nil {
 		return "", fmt.Errorf("asr start: %w", err)
@@ -654,8 +654,8 @@ func (s *voiceState) bargeMonitor(ctx context.Context, cancel context.CancelFunc
 		Channels:         1,
 		SingleSentence:   true,
 		SilenceTimeoutMs: 60000,
-		SpeechTimeoutMs:  s.bot.ASRSpeechTimeoutSec * 1000,
-		SpeechMaxMs:      s.bot.ASRSpeechMaxSec * 1000,
+		SpeechTimeoutMs:  int(s.bot.ASRSpeechTimeoutSec * 1000),
+		SpeechMaxMs:      int(s.bot.ASRSpeechMaxSec * 1000),
 	})
 	if err != nil {
 		s.logger.Debug("barge asr start", "err", err)
